requirements.txt
selenium
webdriver-manager
loguru
pyairtable
python-dotenv
pyperclip